Move the Python-retained value entry onto the bridge behind a registered provider
The retained-value entry -- the ops table for a PyObject retained as the
value itself behind a supported schema, and the policy that says which
schemas may retain, whether an inline cache pays, and how a Python-owned
Bundle binds -- lived inside types/metadata/value_plan_factory.cpp under
build-time Python conditionals. python_bridge.rst named moving it behind a
bridge-registered provider as the remaining layering step for that file.

It now lives on the bridge (src/hgraph/python/impl/retained_value.cpp,
include/hgraph/python/retained_value.h) and reaches the plan factory only
through `ValuePlanFactory::PythonStorageProvider`, a table of plain function
pointers the module registers once at initialisation. The factory carries
no Python type and no build-time conditional for that policy; with no
provider registered every Python-aware storage request downgrades to
Native, so the default Python-free build compiles the same policy code.
`prepare_python_storage_value` is a bridge function; its one type-layer
caller (ts_data_atomic_ops.cpp) calls it there. `type-layer-python-conditionals`
165 -> 160.

Verified: 1696 C++ cases on the default build; 3594 Python tests across
core, extensions and adaptors; the three installed-SDK consumer checks.
